% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cxd_to_avi.R
\name{get_duration_avi}
\alias{get_duration_avi}
\title{Extract duration in seconds using ffmpeg}
\usage{
get_duration_avi(file, ffmpeg = par_ffmpeg(), mc.cores = par_mc.cores())
}
\arguments{
\item{file}{file name, of the video file}

\item{mc.cores}{Number of cores to be used when more than one xd file is given.}
}
\value{
named vector with the duration (in seconds) of the video(s) and the names are the file names
}
\description{
The duration is based on using the output of \code{ffmpeg} and
should therefore work for all ffmpeg supported video formats,
but is only implemented for \code{.avi} files.
}
