% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cxd_to_avi.R
\name{get_fps_avi}
\alias{get_fps_avi}
\title{Extract fps (frames per second) using ffmpeg}
\usage{
get_fps_avi(file, ffmpeg = par_ffmpeg(), mc.cores = par_mc.cores())
}
\arguments{
\item{file}{file name(s), of the video (\code{avi}) file(s) or directory in which the video files are}

\item{mc.cores}{Number of cores to be used when more than one xd file is given.}
}
\value{
named vector with the fps of the video(s) and the names are the file names
}
\description{
The fps is based on using the output of \code{ffmpeg} and
should therefore work for all ffmpeg supported video formats,
but at the moment only supported for \code{avi} files.
}
