% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cxd_to_avi.R
\name{get_fps_cxd}
\alias{get_fps_cxd}
\title{Extract fps (frames per second) from the \code{.cxd} file}
\usage{
get_fps_cxd(file, showinf = par_showinf(), mc.cores = par_mc.cores())
}
\arguments{
\item{file}{file name, of the \code{.cxd} file}
}
\value{
the fps (frames per second) of the video
}
\description{
The fps is based on the results of \code{get_delays_cxd(file)} by averaging these and
calculating from this mean the fps.
}
